+++ integration/spring-data/1.x/src/main/java/com/blazebit/persistence/spring/data/impl/repository/BlazePersistenceRepositoryFactoryBean.java
@@ -118,4 +118,11 @@ public class BlazePersistenceRepositoryFactoryBean<T extends Repository<S, ID>, 
         super.afterPropertiesSet();
     }
 
+    public void setEscapeCharacter(char escapeCharacter) {
+    }
+
+    public char getEscapeCharacter() {
+        return '\\';
+    }
+
 }